modify fault of code
